GUITARModel-Plugin-JFC/tags/03.28.2010/tests/edu/umd/cs/guitar/model/MGUIStructureTest.java
/*
 * To change this template, choose Tools | Templates
 * and open the template in the editor.
 */

package edu.umd.cs.guitar.model;

import java.util.List;
import junit.framework.Test;
import junit.framework.TestCase;
import junit.framework.TestSuite;

/**
 * jUnit tests for MGUIStructure
 */
public class MGUIStructureTest extends TestCase {
    
    public MGUIStructureTest(String testName) {
        super(testName);
    }

    public static Test suite() {
        TestSuite suite = new TestSuite(MGUIStructureTest.class);
        return suite;
    }

    @Override
    protected void setUp() throws Exception {
        super.setUp();
    }

    @Override
    protected void tearDown() throws Exception {
        super.tearDown();
    }

    /**
     * Test of getMRoot method, of class MGUIStructure.
     */
    public void testGetMRoot() {
        System.out.println("getMRoot");
        MGUIStructure instance = null;
        MGUI expResult = null;
        MGUI result = instance.getMRoot();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of getFirstChildByName method, of class MGUIStructure.
     */
    public void testGetFirstChildByName() {
        System.out.println("getFirstChildByName");
        String sName = "";
        MGUIStructure instance = null;
        MComponent expResult = null;
        MComponent result = instance.getFirstChildByName(sName);
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

    /**
     * Test of getLGUI method, of class MGUIStructure.
     */
    public void testGetLGUI() {
        System.out.println("getLGUI");
        MGUIStructure instance = null;
        List expResult = null;
        List result = instance.getLGUI();
        assertEquals(expResult, result);
        // TODO review the generated test code and remove the default call to fail.
        fail("The test case is a prototype.");
    }

}
